Show read-history dialog directly over MainBibleActivity
Long-pressing the chapter check icon previously opened the full
ReadingProgressActivity just to show the history dialog and finish
itself when closed — a heavyweight detour for a modal popup.

Extract the dialog into ReadHistoryDialog as a standalone object that
takes any Activity + LifecycleOwner. Both ReadingProgressActivity and
BibleJavascriptInterface now call ReadHistoryDialog.showFor*() and the
dialog appears in place. DB fetch and bulk delete run on Dispatchers.IO
via lifecycleScope so the binder/main thread is no longer blocked.
Drop the EXTRA_HISTORY_* intent extras and openInitialHistoryDialogIfRequested
that existed only to support the old launch-activity-just-to-popup path.
